Remove RocksDB structs and lib
