+++ server/db-smt/CURRENT
@@ -1,0 +1,1 @@
+MANIFEST-000026
