+++ server/db-smt/OPTIONS-000029
@@ -1,0 +1,164 @@
+# This is a RocksDB option file.
+#
+# For detailed file format spec, please refer to the example file
+# in examples/rocksdb_option_file_example.ini
+#
+
+[Version]
+  rocksdb_version=6.7.3
+  options_file_version=1.1
+
+[DBOptions]
+  manual_wal_flush=false
+  avoid_flush_during_recovery=false
+  dump_malloc_stats=false
+  info_log_level=INFO_LEVEL
+  access_hint_on_compaction_start=NORMAL
+  write_thread_slow_yield_usec=3
+  enable_write_thread_adaptive_yield=true
+  unordered_write=false
+  log_readahead_size=0
+  enable_pipelined_write=false
+  persist_stats_to_disk=false
+  preserve_deletes=false
+  stats_persist_period_sec=600
+  wal_bytes_per_sync=0
+  bytes_per_sync=0
+  delayed_write_rate=16777216
+  WAL_ttl_seconds=0
+  writable_file_max_buffer_size=1048576
+  allow_concurrent_memtable_write=true
+  paranoid_checks=true
+  WAL_size_limit_MB=0
+  write_dbid_to_manifest=false
+  atomic_flush=false
+  max_subcompactions=1
+  wal_dir=../server/db-smt
+  max_total_wal_size=0
+  db_write_buffer_size=0
+  avoid_flush_during_shutdown=false
+  keep_log_file_num=1000
+  table_cache_numshardbits=6
+  max_file_opening_threads=16
+  max_background_compactions=-1
+  two_write_queues=false
+  max_background_jobs=2
+  use_fsync=false
+  random_access_max_buffer_size=1048576
+  max_open_files=-1
+  skip_stats_update_on_db_open=false
+  error_if_exists=false
+  manifest_preallocation_size=4194304
+  max_background_flushes=-1
+  is_fd_close_on_exec=true
+  create_if_missing=true
+  max_write_batch_group_size_bytes=1048576
+  use_adaptive_mutex=false
+  enable_thread_tracking=false
+  stats_dump_period_sec=600
+  max_log_file_size=0
+  advise_random_on_open=true
+  create_missing_column_families=false
+  log_file_time_to_roll=0
+  use_direct_io_for_flush_and_compaction=false
+  compaction_readahead_size=0
+  delete_obsolete_files_period_micros=21600000000
+  recycle_log_file_num=0
+  stats_history_buffer_size=1048576
+  strict_bytes_per_sync=false
+  skip_log_error_on_recovery=false
+  allow_mmap_reads=false
+  allow_ingest_behind=false
+  fail_if_options_file_error=false
+  avoid_unnecessary_blocking_io=false
+  allow_fallocate=true
+  allow_2pc=false
+  max_manifest_file_size=1073741824
+  write_thread_max_yield_usec=100
+  use_direct_reads=false
+  wal_recovery_mode=kPointInTimeRecovery
+  db_log_dir=
+  new_table_reader_for_compaction_inputs=false
+  base_background_compactions=-1
+  allow_mmap_writes=false
+  
+
+[CFOptions "default"]
+  sample_for_compression=0
+  compaction_style=kCompactionStyleLevel
+  merge_operator=nullptr
+  compaction_options_universal={allow_trivial_move=false;stop_style=kCompactionStopStyleTotalSize;min_merge_width=2;compression_size_percent=-1;max_size_amplification_percent=200;max_merge_width=4294967295;size_ratio=1;}
+  compaction_filter=nullptr
+  memtable_factory=SkipListFactory
+  prefix_extractor=nullptr
+  bottommost_compression=kDisableCompressionOption
+  compression_per_level=
+  compression=kSnappyCompression
+  comparator=leveldb.BytewiseComparator
+  max_write_buffer_size_to_maintain=0
+  max_sequential_skip_in_iterations=8
+  max_bytes_for_level_base=268435456
+  memtable_insert_with_hint_prefix_extractor=nullptr
+  memtable_whole_key_filtering=false
+  soft_pending_compaction_bytes_limit=68719476736
+  compaction_pri=kMinOverlappingRatio
+  memtable_huge_page_size=0
+  max_successive_merges=0
+  min_write_buffer_number_to_merge=1
+  target_file_size_multiplier=1
+  table_factory=BlockBasedTable
+  num_levels=7
+  arena_block_size=8388608
+  periodic_compaction_seconds=0
+  level0_stop_writes_trigger=36
+  level0_file_num_compaction_trigger=4
+  inplace_update_num_locks=10000
+  compaction_options_fifo={allow_compaction=false;max_table_files_size=1073741824;}
+  level_compaction_dynamic_level_bytes=false
+  level0_slowdown_writes_trigger=20
+  max_bytes_for_level_multiplier_additional=1:1:1:1:1:1:1
+  ttl=2592000
+  max_write_buffer_number=2
+  force_consistency_checks=false
+  memtable_prefix_bloom_size_ratio=0.000000
+  paranoid_file_checks=false
+  hard_pending_compaction_bytes_limit=274877906944
+  max_write_buffer_number_to_maintain=0
+  target_file_size_base=67108864
+  optimize_filters_for_hits=false
+  compaction_filter_factory=nullptr
+  max_bytes_for_level_multiplier=10.000000
+  bloom_locality=0
+  write_buffer_size=67108864
+  disable_auto_compactions=false
+  max_compaction_bytes=1677721600
+  inplace_update_support=false
+  report_bg_io_stats=false
+  
+[TableOptions/BlockBasedTable "default"]
+  read_amp_bytes_per_bit=8589934592
+  format_version=2
+  verify_compression=false
+  partition_filters=false
+  enable_index_compression=true
+  checksum=kCRC32c
+  index_block_restart_interval=1
+  pin_top_level_index_and_filter=true
+  block_align=false
+  block_size=4096
+  index_type=kBinarySearch
+  filter_policy=nullptr
+  metadata_block_size=4096
+  no_block_cache=false
+  whole_key_filtering=true
+  index_shortening=kShortenSeparators
+  block_size_deviation=10
+  data_block_index_type=kDataBlockBinarySearch
+  data_block_hash_table_util_ratio=0.750000
+  cache_index_and_filter_blocks=false
+  block_restart_interval=16
+  pin_l0_filter_and_index_blocks_in_cache=false
+  hash_index_allow_collision=true
+  cache_index_and_filter_blocks_with_high_priority=true
+  flush_block_policy_factory=FlushBlockBySizePolicyFactory
+
