+++ server/src/storage/mod.rs
@@ -1,2 +1,1 @@
 pub mod db;
-pub mod db_postgres;
